+++ apps/runtime-demo/3006-runtime-remote/webpack.config.js
@@ -12,7 +12,6 @@ module.exports = composePlugins(
   withNx(),
   withReact(),
   async (config, context) => {
-    // const webpack = require('/Users/bytedance/outter/universe/node_modules/.pnpm/webpack@5.89.0_@swc+core@1.3.99_esbuild@0.19.7/node_modules/webpack/lib/index.js')
     // const ModuleFederationPlugin = webpack.container.ModuleFederationPlugin;
     config.plugins.push(
       new ModuleFederationPlugin({