+++ packages/enhanced/src/schemas/container/ContainerPlugin.check.ts
@@ -1899,8 +1899,7 @@ function validate19(
   data,
   { instancePath = '', parentData, parentDataProperty, rootData = data } = {},
 ) {
-  /*# sourceURL="file:///Users/bytedance/work/webpack/schemas/plugins/container/ContainerPlugin.json" */ let vErrors =
-    null;
+  /*# sourceURL="" */ let vErrors = null;
   let errors = 0;
   if (errors === 0) {
     if (data && typeof data == 'object' && !Array.isArray(data)) {