+++ packages/managers/src/index.ts
@@ -1,0 +1,7 @@
+export { BasicPluginOptionsManager } from './BasicPluginOptionsManager';
+export { ContainerManager } from './ContainerManager';
+export { PKGJsonManager } from './PKGJsonManager';
+export { RemoteManager } from './RemoteManager';
+export { SharedManager } from './SharedManager';
+export * as utils from './utils';
+export * as types from './types';