+++ packages/native-federation-tests/src/index.test.ts
@@ -95,6 +95,49 @@ describe('index', () => {
         },
       });
     });
+
+    it('correctly enrich rspack config', async () => {
+      const options = {
+        moduleFederationConfig: {
+          name: 'moduleFederationTypescript',
+          filename: 'remoteEntry.js',
+          exposes: {
+            './index': exposedIndex,
+          },
+          shared: {
+            react: { singleton: true, eager: true },
+            'react-dom': { singleton: true, eager: true },
+          },
+        },
+        deleteTestsFolder: false,
+        testsFolder: '@mf-tests',
+      };
+
+      const rspackCompiler = {
+        options: {
+          devServer: {
+            foo: {},
+          },
+        },
+      } as any;
+
+      const unplugin = NativeFederationTestsRemote.rollup(
+        options,
+      ) as UnpluginOptions;
+
+      unplugin.rspack?.(rspackCompiler);
+
+      expect(rspackCompiler).toStrictEqual({
+        options: {
+          devServer: {
+            foo: {},
+            static: {
+              directory: resolve('./dist'),
+            },
+          },
+        },
+      });
+    });
   });
 
   describe('NativeFederationTestsHost', () => {
@@ -132,7 +175,9 @@ describe('index', () => {
       ) as UnpluginOptions;
       await expect(unplugin.writeBundle?.()).resolves.not.toThrow();
 
-      expect(dirTree(options.mocksFolder)).toMatchObject({
+      const testsFolder = join(projectRoot, options.mocksFolder);
+
+      expect(dirTree(testsFolder)).toMatchObject({
         name: '__mocks__',
         children: [
           {