+++ packages/runtime/__tests__/resources/register-remotes/app2/federation-remote-entry.js
@@ -1,0 +1,11 @@
+globalThis[`@register-remotes/app2`] = {
+  get(scope) {
+    const moduleMap = {
+      './say'() {
+        return () => 'hello app2';
+      },
+    };
+    return moduleMap[scope];
+  },
+  init() {},
+};