+++ packages/runtime/src/index.ts
@@ -10,7 +10,7 @@ import { assert } from './utils/logger';
 
 export { FederationHost } from './core';
 export { registerGlobalPlugins } from './global';
-export { loadScript } from '@module-federation/sdk';
+export { loadScript, loadScriptNode } from '@module-federation/sdk';
 
 export type { Federation } from './global';
 export type { FederationRuntimePlugin };
@@ -66,6 +66,14 @@ export function preloadRemote(
   assert(FederationInstance, 'Please call init first');
   // eslint-disable-next-line prefer-spread
   return FederationInstance.preloadRemote.apply(FederationInstance, args);
+}
+
+export function registerRemotes(
+  ...args: Parameters<FederationHost['registerRemotes']>
+): ReturnType<FederationHost['registerRemotes']> {
+  assert(FederationInstance, 'Please call init first');
+  // eslint-disable-next-line prefer-spread
+  return FederationInstance.registerRemotes.apply(FederationInstance, args);
 }
 
 // Inject for debug