+++ packages/runtime/src/utils/load.ts
@@ -20,9 +20,9 @@ export async function loadEsmEntry({
       if (!remoteEntryExports) {
         // eslint-disable-next-line no-eval
         new Function(
-          'resolve',
-          `import("${entry}").then((res)=>{resolve(res);}, (error)=> reject(error))`,
-        )(resolve);
+          'callbacks',
+          `import("${entry}").then(callbacks[0]).catch(callbacks[1])`,
+        )([resolve, reject]);
       } else {
         resolve(remoteEntryExports);
       }
@@ -75,7 +75,6 @@ export async function loadEntryScript({
         2. ${remoteEntryKey} cannot be used to get remoteEntry exports in the window object.
       `,
         );
-        console.log(entryExports);
 
         return entryExports;
       })
@@ -108,6 +107,11 @@ export async function loadEntryScript({
     });
 }
 
+export function getRemoteEntryUniqueKey(remoteInfo: RemoteInfo): string {
+  const { entry, name } = remoteInfo;
+  return composeKeyWithSeparator(name, entry);
+}
+
 export async function getRemoteEntry({
   remoteEntryExports,
   remoteInfo,
@@ -120,7 +124,7 @@ export async function getRemoteEntry({
   fetchHook?: (input: string, init?: RequestInit) => Promise<Response | void | false>;
 }): Promise<RemoteEntryExports | void> {
   const { entry, name, type, entryGlobalName } = remoteInfo;
-  const uniqueKey = composeKeyWithSeparator(name, entry);
+  const uniqueKey = getRemoteEntryUniqueKey(remoteInfo);
   if (remoteEntryExports) {
     return remoteEntryExports;
   }