+++ packages/enhanced/CHANGELOG.md
@@ -1,5 +1,69 @@
 # [0.2.0-canary.5](https://github.com/module-federation/universe/compare/enhanced-0.2.0-canary.4...enhanced-0.2.0-canary.5) (2023-11-20)
 
+## 0.0.17
+
+### Patch Changes
+
+- @module-federation/runtime-tools@0.0.17
+- @module-federation/sdk@0.0.17
+- @module-federation/managers@0.0.17
+- @module-federation/manifest@0.0.17
+
+## 0.0.16
+
+### Patch Changes
+
+- Updated dependencies [ccafac3]
+  - @module-federation/manifest@0.0.16
+  - @module-federation/sdk@0.0.16
+  - @module-federation/runtime-tools@0.0.16
+  - @module-federation/managers@0.0.16
+
+## 0.0.15
+
+### Patch Changes
+
+- ba5bedd: feat: support manifest
+- Updated dependencies [3a45d99]
+- Updated dependencies [ba5bedd]
+  - @module-federation/sdk@0.0.15
+  - @module-federation/managers@0.0.15
+  - @module-federation/manifest@0.0.15
+  - @module-federation/runtime-tools@0.0.15
+
+## 0.0.14
+
+### Patch Changes
+
+- Updated dependencies [cd8c7bf]
+- Updated dependencies [5576c6b]
+  - @module-federation/sdk@0.0.14
+  - @module-federation/runtime-tools@0.0.14
+
+## 0.0.13
+
+### Patch Changes
+
+- Updated dependencies [804447c]
+  - @module-federation/sdk@0.0.13
+  - @module-federation/runtime-tools@0.0.13
+
+## 0.0.12
+
+### Patch Changes
+
+- 4ca0c7b: fix(enhanced): use upath to against more edge cases between different OS
+  - @module-federation/sdk@0.0.12
+  - @module-federation/runtime-tools@0.0.12
+
+## 0.0.11
+
+### Patch Changes
+
+- 5c17bc4: normalize bundler runtime import paths
+  - @module-federation/runtime-tools@0.0.11
+  - @module-federation/sdk@0.0.11
+
 ## 0.0.10
 
 ### Patch Changes